Refactoring the "PortfolioViewerWidget" class
Refactoring the "PortfolioViewerWidget" class, in order to split it in more files
